+++ src/main/java/com/depromeet/warmup1/dto/LoginDto.java
@@ -1,0 +1,8 @@
+package com.depromeet.warmup1.dto;
+
+import lombok.Data;
+
+@Data
+public class LoginDto {
+	private String kakaoToken;
+}
