+++ src/main/java/com/depromeet/warmup1/entity/Member.java
@@ -24,5 +24,5 @@ public class Member {
 	private String kakaoId;
 	private String name;
 	private String profileImgUrl;
-	private String firebaseToken;
+	private String connectKey;
 }
